Fixing warning about duplicate key
This may be a Ruby version difference? It now complains
about the duplicate key "issue_checklist", which is on
this object twice.

The API in question does let you pass an array as the
value for the "issue_checklist" field, so the right
thing to do here is move the two values into an array,
instead of repeating the key.